+++ app_store_client/__manifest__.py
@@ -1,6 +1,6 @@
 {
     'name': "Custom App Store Client",
-    'version': "1.0.4",
+    'version': "1.0.5",
     'author': "Sythil Tech",
     'category': "Tools",
     'support': "steven@sythiltech.com.au",