+++ voip_sip_webrtc_twilio/__manifest__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 {
     'name': "Voip Communication - Twilio",
-    'version': "1.0.22",
+    'version': "1.0.23",
     'author': "Sythil Tech",
     'category': "Tools",
     'support': "steven@sythiltech.com.au",
@@ -19,6 +19,7 @@
         'views/voip_account_action_views.xml',
         'views/voip_sip_webrtc_twilio_templates.xml',
         'views/voip_call_comment_views.xml',
+        'views/mail_activity_views.xml',
         'views/menus.xml',
         #'data/voip.account.action.type.csv',
         'security/ir.model.access.csv',