+++ voip_sip_webrtc_twilio/models/__init__.py
@@ -7,3 +7,4 @@ from . import res_partner
 from . import voip_number
 from . import voip_account_action
 from . import crm_lead
+from . import mail_activity